Make StyleBox use int instead of float for margin/grow

+++ scene/resources/style_box.cpp
@@ -38,17 +38,17 @@ bool StyleBox::test_mask(const Point2 &p_point, const Rect2 &p_rect) const {
 	return true;
 }
 
-void StyleBox::set_default_margin(Margin p_margin, float p_value) {
+void StyleBox::set_default_margin(Margin p_margin, int p_value) {
 
 	margin[p_margin] = p_value;
 	emit_changed();
 }
-float StyleBox::get_default_margin(Margin p_margin) const {
+int StyleBox::get_default_margin(Margin p_margin) const {
 
 	return margin[p_margin];
 }
 
-float StyleBox::get_margin(Margin p_margin) const {
+int StyleBox::get_margin(Margin p_margin) const {
 
 	if (margin[p_margin] < 0)
 		return get_style_margin(p_margin);
@@ -94,10 +94,10 @@ void StyleBox::_bind_methods() {
 	ClassDB::bind_method(D_METHOD("draw", "canvas_item", "rect"), &StyleBox::draw);
 
 	ADD_GROUP("Content Margin", "content_margin_");
-	ADD_PROPERTYI(PropertyInfo(Variant::REAL, "content_margin_left", PROPERTY_HINT_RANGE, "-1,2048,1"), "set_default_margin", "get_default_margin", MARGIN_LEFT);
-	ADD_PROPERTYI(PropertyInfo(Variant::REAL, "content_margin_right", PROPERTY_HINT_RANGE, "-1,2048,1"), "set_default_margin", "get_default_margin", MARGIN_RIGHT);
-	ADD_PROPERTYI(PropertyInfo(Variant::REAL, "content_margin_top", PROPERTY_HINT_RANGE, "-1,2048,1"), "set_default_margin", "get_default_margin", MARGIN_TOP);
-	ADD_PROPERTYI(PropertyInfo(Variant::REAL, "content_margin_bottom", PROPERTY_HINT_RANGE, "-1,2048,1"), "set_default_margin", "get_default_margin", MARGIN_BOTTOM);
+	ADD_PROPERTYI(PropertyInfo(Variant::INT, "content_margin_left", PROPERTY_HINT_RANGE, "-1,2048,1"), "set_default_margin", "get_default_margin", MARGIN_LEFT);
+	ADD_PROPERTYI(PropertyInfo(Variant::INT, "content_margin_right", PROPERTY_HINT_RANGE, "-1,2048,1"), "set_default_margin", "get_default_margin", MARGIN_RIGHT);
+	ADD_PROPERTYI(PropertyInfo(Variant::INT, "content_margin_top", PROPERTY_HINT_RANGE, "-1,2048,1"), "set_default_margin", "get_default_margin", MARGIN_TOP);
+	ADD_PROPERTYI(PropertyInfo(Variant::INT, "content_margin_bottom", PROPERTY_HINT_RANGE, "-1,2048,1"), "set_default_margin", "get_default_margin", MARGIN_BOTTOM);
 }
 
 StyleBox::StyleBox() {
@@ -141,7 +141,7 @@ Ref<Texture> StyleBoxTexture::get_normal_map() const {
 	return normal_map;
 }
 
-void StyleBoxTexture::set_margin_size(Margin p_margin, float p_size) {
+void StyleBoxTexture::set_margin_size(Margin p_margin, int p_size) {
 
 	ERR_FAIL_INDEX((int)p_margin, 4);
 
@@ -155,12 +155,12 @@ void StyleBoxTexture::set_margin_size(Margin p_margin, float p_size) {
 	};
 	_change_notify(margin_prop[p_margin]);
 }
-float StyleBoxTexture::get_margin_size(Margin p_margin) const {
+int StyleBoxTexture::get_margin_size(Margin p_margin) const {
 
 	return margin[p_margin];
 }
 
-float StyleBoxTexture::get_style_margin(Margin p_margin) const {
+int StyleBoxTexture::get_style_margin(Margin p_margin) const {
 
 	return margin[p_margin];
 }
@@ -205,14 +205,14 @@ Size2 StyleBoxTexture::get_center_size() const {
 	return region_rect.size - get_minimum_size();
 }
 
-void StyleBoxTexture::set_expand_margin_size(Margin p_expand_margin, float p_size) {
+void StyleBoxTexture::set_expand_margin_size(Margin p_expand_margin, int p_size) {
 
 	ERR_FAIL_INDEX((int)p_expand_margin, 4);
 	expand_margin[p_expand_margin] = p_size;
 	emit_changed();
 }
 
-void StyleBoxTexture::set_expand_margin_size_individual(float p_left, float p_top, float p_right, float p_bottom) {
+void StyleBoxTexture::set_expand_margin_size_individual(int p_left, int p_top, int p_right, int p_bottom) {
 	expand_margin[MARGIN_LEFT] = p_left;
 	expand_margin[MARGIN_TOP] = p_top;
 	expand_margin[MARGIN_RIGHT] = p_right;
@@ -220,7 +220,7 @@ void StyleBoxTexture::set_expand_margin_size_individual(float p_left, float p_to
 	emit_changed();
 }
 
-void StyleBoxTexture::set_expand_margin_size_all(float p_expand_margin_size) {
+void StyleBoxTexture::set_expand_margin_size_all(int p_expand_margin_size) {
 	for (int i = 0; i < 4; i++) {
 
 		expand_margin[i] = p_expand_margin_size;
@@ -228,7 +228,7 @@ void StyleBoxTexture::set_expand_margin_size_all(float p_expand_margin_size) {
 	emit_changed();
 }
 
-float StyleBoxTexture::get_expand_margin_size(Margin p_expand_margin) const {
+int StyleBoxTexture::get_expand_margin_size(Margin p_expand_margin) const {
 
 	ERR_FAIL_INDEX_V((int)p_expand_margin, 4, 0);
 	return expand_margin[p_expand_margin];
@@ -319,15 +319,15 @@ void StyleBoxTexture::_bind_methods() {
 	ADD_PROPERTY(PropertyInfo(Variant::OBJECT, "normal_map", PROPERTY_HINT_RESOURCE_TYPE, "Texture"), "set_normal_map", "get_normal_map");
 	ADD_PROPERTY(PropertyInfo(Variant::RECT2, "region_rect"), "set_region_rect", "get_region_rect");
 	ADD_GROUP("Margin", "margin_");
-	ADD_PROPERTYI(PropertyInfo(Variant::REAL, "margin_left", PROPERTY_HINT_RANGE, "0,2048,1"), "set_margin_size", "get_margin_size", MARGIN_LEFT);
-	ADD_PROPERTYI(PropertyInfo(Variant::REAL, "margin_right", PROPERTY_HINT_RANGE, "0,2048,1"), "set_margin_size", "get_margin_size", MARGIN_RIGHT);
-	ADD_PROPERTYI(PropertyInfo(Variant::REAL, "margin_top", PROPERTY_HINT_RANGE, "0,2048,1"), "set_margin_size", "get_margin_size", MARGIN_TOP);
-	ADD_PROPERTYI(PropertyInfo(Variant::REAL, "margin_bottom", PROPERTY_HINT_RANGE, "0,2048,1"), "set_margin_size", "get_margin_size", MARGIN_BOTTOM);
+	ADD_PROPERTYI(PropertyInfo(Variant::INT, "margin_left", PROPERTY_HINT_RANGE, "0,2048,1"), "set_margin_size", "get_margin_size", MARGIN_LEFT);
+	ADD_PROPERTYI(PropertyInfo(Variant::INT, "margin_right", PROPERTY_HINT_RANGE, "0,2048,1"), "set_margin_size", "get_margin_size", MARGIN_RIGHT);
+	ADD_PROPERTYI(PropertyInfo(Variant::INT, "margin_top", PROPERTY_HINT_RANGE, "0,2048,1"), "set_margin_size", "get_margin_size", MARGIN_TOP);
+	ADD_PROPERTYI(PropertyInfo(Variant::INT, "margin_bottom", PROPERTY_HINT_RANGE, "0,2048,1"), "set_margin_size", "get_margin_size", MARGIN_BOTTOM);
 	ADD_GROUP("Expand Margin", "expand_margin_");
-	ADD_PROPERTYI(PropertyInfo(Variant::REAL, "expand_margin_left", PROPERTY_HINT_RANGE, "0,2048,1"), "set_expand_margin_size", "get_expand_margin_size", MARGIN_LEFT);
-	ADD_PROPERTYI(PropertyInfo(Variant::REAL, "expand_margin_right", PROPERTY_HINT_RANGE, "0,2048,1"), "set_expand_margin_size", "get_expand_margin_size", MARGIN_RIGHT);
-	ADD_PROPERTYI(PropertyInfo(Variant::REAL, "expand_margin_top", PROPERTY_HINT_RANGE, "0,2048,1"), "set_expand_margin_size", "get_expand_margin_size", MARGIN_TOP);
-	ADD_PROPERTYI(PropertyInfo(Variant::REAL, "expand_margin_bottom", PROPERTY_HINT_RANGE, "0,2048,1"), "set_expand_margin_size", "get_expand_margin_size", MARGIN_BOTTOM);
+	ADD_PROPERTYI(PropertyInfo(Variant::INT, "expand_margin_left", PROPERTY_HINT_RANGE, "0,2048,1"), "set_expand_margin_size", "get_expand_margin_size", MARGIN_LEFT);
+	ADD_PROPERTYI(PropertyInfo(Variant::INT, "expand_margin_right", PROPERTY_HINT_RANGE, "0,2048,1"), "set_expand_margin_size", "get_expand_margin_size", MARGIN_RIGHT);
+	ADD_PROPERTYI(PropertyInfo(Variant::INT, "expand_margin_top", PROPERTY_HINT_RANGE, "0,2048,1"), "set_expand_margin_size", "get_expand_margin_size", MARGIN_TOP);
+	ADD_PROPERTYI(PropertyInfo(Variant::INT, "expand_margin_bottom", PROPERTY_HINT_RANGE, "0,2048,1"), "set_expand_margin_size", "get_expand_margin_size", MARGIN_BOTTOM);
 	ADD_GROUP("Axis Stretch", "axis_stretch_");
 	ADD_PROPERTY(PropertyInfo(Variant::INT, "axis_stretch_horizontal", PROPERTY_HINT_ENUM, "Stretch,Tile,Tile Fit"), "set_h_axis_stretch_mode", "get_h_axis_stretch_mode");
 	ADD_PROPERTY(PropertyInfo(Variant::INT, "axis_stretch_vertical", PROPERTY_HINT_ENUM, "Stretch,Tile,Tile Fit"), "set_v_axis_stretch_mode", "get_v_axis_stretch_mode");
@@ -455,13 +455,13 @@ int StyleBoxFlat::get_corner_radius(const Corner p_corner) const {
 	return corner_radius[p_corner];
 }
 
-void StyleBoxFlat::set_expand_margin_size(Margin p_expand_margin, float p_size) {
+void StyleBoxFlat::set_expand_margin_size(Margin p_expand_margin, int p_size) {
 
 	expand_margin[p_expand_margin] = p_size;
 	emit_changed();
 }
 
-void StyleBoxFlat::set_expand_margin_size_individual(float p_left, float p_top, float p_right, float p_bottom) {
+void StyleBoxFlat::set_expand_margin_size_individual(int p_left, int p_top, int p_right, int p_bottom) {
 	expand_margin[MARGIN_LEFT] = p_left;
 	expand_margin[MARGIN_TOP] = p_top;
 	expand_margin[MARGIN_RIGHT] = p_right;
@@ -469,7 +469,7 @@ void StyleBoxFlat::set_expand_margin_size_individual(float p_left, float p_top, 
 	emit_changed();
 }
 
-void StyleBoxFlat::set_expand_margin_size_all(float p_expand_margin_size) {
+void StyleBoxFlat::set_expand_margin_size_all(int p_expand_margin_size) {
 	for (int i = 0; i < 4; i++) {
 
 		expand_margin[i] = p_expand_margin_size;
@@ -477,7 +477,7 @@ void StyleBoxFlat::set_expand_margin_size_all(float p_expand_margin_size) {
 	emit_changed();
 }
 
-float StyleBoxFlat::get_expand_margin_size(Margin p_expand_margin) const {
+int StyleBoxFlat::get_expand_margin_size(Margin p_expand_margin) const {
 
 	return expand_margin[p_expand_margin];
 }
@@ -762,7 +762,7 @@ void StyleBoxFlat::draw(RID p_canvas_item, const Rect2 &p_rect) const {
 	vs->canvas_item_add_triangle_array(p_canvas_item, indices, verts, colors);
 }
 
-float StyleBoxFlat::get_style_margin(Margin p_margin) const {
+int StyleBoxFlat::get_style_margin(Margin p_margin) const {
 	return border_width[p_margin];
 }
 void StyleBoxFlat::_bind_methods() {
@@ -835,10 +835,10 @@ void StyleBoxFlat::_bind_methods() {
 	ADD_PROPERTY(PropertyInfo(Variant::INT, "corner_detail", PROPERTY_HINT_RANGE, "1,128,1"), "set_corner_detail", "get_corner_detail");
 
 	ADD_GROUP("Expand Margin", "expand_margin_");
-	ADD_PROPERTYI(PropertyInfo(Variant::REAL, "expand_margin_left", PROPERTY_HINT_RANGE, "0,2048,1"), "set_expand_margin", "get_expand_margin", MARGIN_LEFT);
-	ADD_PROPERTYI(PropertyInfo(Variant::REAL, "expand_margin_right", PROPERTY_HINT_RANGE, "0,2048,1"), "set_expand_margin", "get_expand_margin", MARGIN_RIGHT);
-	ADD_PROPERTYI(PropertyInfo(Variant::REAL, "expand_margin_top", PROPERTY_HINT_RANGE, "0,2048,1"), "set_expand_margin", "get_expand_margin", MARGIN_TOP);
-	ADD_PROPERTYI(PropertyInfo(Variant::REAL, "expand_margin_bottom", PROPERTY_HINT_RANGE, "0,2048,1"), "set_expand_margin", "get_expand_margin", MARGIN_BOTTOM);
+	ADD_PROPERTYI(PropertyInfo(Variant::INT, "expand_margin_left", PROPERTY_HINT_RANGE, "0,2048,1"), "set_expand_margin", "get_expand_margin", MARGIN_LEFT);
+	ADD_PROPERTYI(PropertyInfo(Variant::INT, "expand_margin_right", PROPERTY_HINT_RANGE, "0,2048,1"), "set_expand_margin", "get_expand_margin", MARGIN_RIGHT);
+	ADD_PROPERTYI(PropertyInfo(Variant::INT, "expand_margin_top", PROPERTY_HINT_RANGE, "0,2048,1"), "set_expand_margin", "get_expand_margin", MARGIN_TOP);
+	ADD_PROPERTYI(PropertyInfo(Variant::INT, "expand_margin_bottom", PROPERTY_HINT_RANGE, "0,2048,1"), "set_expand_margin", "get_expand_margin", MARGIN_BOTTOM);
 
 	ADD_GROUP("Shadow", "shadow_");
 	ADD_PROPERTY(PropertyInfo(Variant::COLOR, "shadow_color"), "set_shadow_color", "get_shadow_color");
@@ -909,19 +909,19 @@ bool StyleBoxLine::is_vertical() const {
 	return vertical;
 }
 
-void StyleBoxLine::set_grow_end(float p_grow_end) {
+void StyleBoxLine::set_grow_end(int p_grow_end) {
 	grow_end = p_grow_end;
 	emit_changed();
 }
-float StyleBoxLine::get_grow_end() const {
+int StyleBoxLine::get_grow_end() const {
 	return grow_end;
 }
 
-void StyleBoxLine::set_grow_begin(float p_grow_begin) {
+void StyleBoxLine::set_grow_begin(int p_grow_begin) {
 	grow_begin = p_grow_begin;
 	emit_changed();
 }
-float StyleBoxLine::get_grow_begin() const {
+int StyleBoxLine::get_grow_begin() const {
 	return grow_begin;
 }
 
@@ -939,12 +939,12 @@ void StyleBoxLine::_bind_methods() {
 	ClassDB::bind_method(D_METHOD("is_vertical"), &StyleBoxLine::is_vertical);
 
 	ADD_PROPERTY(PropertyInfo(Variant::COLOR, "color"), "set_color", "get_color");
-	ADD_PROPERTY(PropertyInfo(Variant::REAL, "grow_begin", PROPERTY_HINT_RANGE, "-300,300,1"), "set_grow_begin", "get_grow_begin");
-	ADD_PROPERTY(PropertyInfo(Variant::REAL, "grow_end", PROPERTY_HINT_RANGE, "-300,300,1"), "set_grow_end", "get_grow_end");
+	ADD_PROPERTY(PropertyInfo(Variant::INT, "grow_begin", PROPERTY_HINT_RANGE, "-300,300,1"), "set_grow_begin", "get_grow_begin");
+	ADD_PROPERTY(PropertyInfo(Variant::INT, "grow_end", PROPERTY_HINT_RANGE, "-300,300,1"), "set_grow_end", "get_grow_end");
 	ADD_PROPERTY(PropertyInfo(Variant::INT, "thickness", PROPERTY_HINT_RANGE, "0,10"), "set_thickness", "get_thickness");
 	ADD_PROPERTY(PropertyInfo(Variant::BOOL, "vertical"), "set_vertical", "is_vertical");
 }
-float StyleBoxLine::get_style_margin(Margin p_margin) const {
+int StyleBoxLine::get_style_margin(Margin p_margin) const {
 	return thickness;
 }
 Size2 StyleBoxLine::get_center_size() const {
